// generated/api/admin/models/box.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * BDUI API
 * Спецификация OpenAPI для Backend-Driven UI  приложения

 * OpenAPI spec version: 1.0.0
 */
import type { CompositeBase } from './compositeBase';
import type { BoxAllOf } from './boxAllOf';
import type { BoxType } from './boxType';

export type Box = CompositeBase &
  BoxAllOf & {
    type: BoxType;
  };