// generated/api/admin/models/dynamicColumn.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * BDUI API
 * Спецификация OpenAPI для Backend-Driven UI  приложения

 * OpenAPI spec version: 1.0.0
 */
import type { DynamicCompositeBase } from './dynamicCompositeBase';
import type { DynamicColumnAllOf } from './dynamicColumnAllOf';
import type { DynamicColumnType } from './dynamicColumnType';

export type DynamicColumn = DynamicCompositeBase &
  DynamicColumnAllOf & {
    type: DynamicColumnType;
  };